refactor(materializations): move formatters to shared utils directory
